Add files via upload
This code has an improved version of threading and camera management. It allows for n number of cameras to be easily added, even though we will only use 2. This also helps with the lag and video formatting. It may take a few seconds for the cameras to first open depending on the camera being used. My Logitech Webcam took about 30 seconds before it opened up.